rename interface and class uses
