+++ src/components/CourseList.vue
@@ -14,8 +14,8 @@ import LecturerCourse from "./LecturerCourse.vue";
 import StudentCourse from "./StudentCourse.vue";
 import Course_Management from "@/api/Course_Management";
 // eslint-disable-next-line no-unused-vars
-import { Course } from "@/entities/Course"
-import ICourse from '@/api/api_models/course_management/ICourse';
+import { CourseEntity } from "@/entities/CourseEntity"
+import Course from '@/api/api_models/course_management/Course';
 
 export default {
     name: "CourseList",
@@ -25,7 +25,7 @@ export default {
     },
     
     async setup() {
-        let courses: ICourse[] = [];
+        let courses: Course[] = [];
         let role = store.state.myRole;
         let roles = Object.values(Role).filter(e => e != Role.NONE);
         let isLecturer: boolean = (role == Role.LECTURER);
@@ -33,7 +33,7 @@ export default {
         let course_management: Course_Management = new Course_Management();
         let myId = store.state.myId;
 
-        await course_management.getCourses().then((response : ICourse[]) => {
+        await course_management.getCourses().then((response : Course[]) => {
             courses = response;
         })
